requirements.txt
Flask==3.0.0
pandas==2.2.0
numpy==1.26.0
reportlab==4.0.7
openpyxl==3.1.5
Werkzeug==3.0.0
python-docx==1.1.0
python-dotenv==1.0.1
# Optional LLM libraries (installed for text formatting)
openai==1.12.0
# anthropic==0.18.1  # Alternative LLM provider
